requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
